Added class Timer and change delta type to double

+++ src/Game/GameObjects/BetonWall.cpp
@@ -63,6 +63,6 @@ void BetonWall::render() const
 	renderBlock(EBlockLocation::TopRight);
 }
 
-void BetonWall::update(const uint64_t delta)
+void BetonWall::update(const double delta)
 {
 }
