+++ src/Game/Level.h
@@ -13,16 +13,16 @@ public:
 	static constexpr unsigned int BLOCK_SIZE = 16;
 	Level(const std::vector<std::string>& levelDescription);
 	void render() const;
-	void update(const uint64_t delta);
+	void update(const double delta);
 	size_t getLevelWidth() const;
 	size_t getLevelHeight() const;
 
-	const glm::vec2& getPlayerRespawn_1() const { return m_playerRespawn_1; }
-	const glm::vec2& getPlayerRespawn_2() const { return m_playerRespawn_2; }
+	const glm::ivec2& getPlayerRespawn_1() const { return m_playerRespawn_1; }
+	const glm::ivec2& getPlayerRespawn_2() const { return m_playerRespawn_2; }
 
-	const glm::vec2& getEnemyRespawn_1() const { return m_enemyRespawn_1; }
-	const glm::vec2& getEnemyRespawn_2() const { return m_enemyRespawn_2; }
-	const glm::vec2& getEnemyRespawn_3() const { return m_enemyRespawn_3; }
+	const glm::ivec2& getEnemyRespawn_1() const { return m_enemyRespawn_1; }
+	const glm::ivec2& getEnemyRespawn_2() const { return m_enemyRespawn_2; }
+	const glm::ivec2& getEnemyRespawn_3() const { return m_enemyRespawn_3; }
 
 private:
 	size_t m_width;
