+++ src/Resources/ResourceManager.cpp
@@ -268,7 +268,7 @@ bool ResourceManager::loadJSONResources(const std::string& JSONPath)
 				for (const auto& currentFrame : framesArray)
 				{
 					const std::string subTexture = currentFrame["subTexture"].GetString();
-					const uint64_t duration = currentFrame["duration"].GetUint64();
+					const double duration = currentFrame["duration"].GetDouble();
 					const auto pTextureAtlas = getTexture(textureAtlas);
 					const auto pSubTexture = pTextureAtlas->getSubTexture(subTexture);
